replace environment view with header in fullscreen view
